requirements.txt
requests
pandas
pymupdf #fitz
ufal-morphodita